Login: Delete unnecessary files
