+++ src/components/Auth/AuthContent/AuthContent.component.tsx
@@ -6,8 +6,8 @@ import { FaCode } from "react-icons/fa";
 type AuthMode = "login" | "register" | "forgot";
 
 const LoginForm = lazy(() => import("../Forms/LoginForm/LoginForm.component"));
-const RegisterForm = lazy(() => import("./RegisterForm.component"));
-const ForgotPasswordForm = lazy(() => import("./ForgotPasswordForm.component"));
+const RegisterForm = lazy(() => import("../Forms/RegisterForm/RegisterForm.component"));
+const ForgotPasswordForm = lazy(() => import("../Forms/ForgotPasswordForm/ForgotPasswordForm.component"));
 
 export const AuthContent = () => {
   const [authMode, setAuthMode] = useState<AuthMode>("login");
