+++ src/components/Auth/Forms/RegisterForm/RegisterForm.component.tsx
@@ -1,32 +1,13 @@
 import { motion } from "framer-motion";
 import { FaEnvelope, FaLock, FaUser } from "react-icons/fa";
-import { useForm } from "react-hook-form";
-import { zodResolver } from "@hookform/resolvers/zod";
 
+import { FormInput } from "../../../UI/Form/FormInput.component";
 import { Button } from "../../../UI/Button/Button.component";
-import { FormInput } from "../../../UI/Form/FormInput/FormInput.component";
-import { RegisterFormData, registerSchema } from "../../../../schemas/auth.schema";
 
-export const RegisterForm = () => {
-  const {
-    register,
-    handleSubmit,
-    formState: { errors, isSubmitting },
-  } = useForm<RegisterFormData>({
-    resolver: zodResolver(registerSchema),
-  });
 
-  const onSubmit = async (data: RegisterFormData) => {
-    try {
-      console.log(data); // Tutaj będzie integracja z API
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
+ const RegisterForm = () => {
   return (
     <motion.form
-      onSubmit={handleSubmit(onSubmit)}
       initial={{ opacity: 0, x: 20 }}
       animate={{ opacity: 1, x: 0 }}
       exit={{ opacity: 0, x: -20 }}
@@ -38,8 +19,6 @@ export const RegisterForm = () => {
         label="Nazwa użytkownika"
         placeholder="jankowalski"
         icon={<FaUser />}
-        error={errors.username?.message}
-        {...register("username")}
       />
 
       <FormInput
@@ -47,8 +26,6 @@ export const RegisterForm = () => {
         label="Email"
         placeholder="twoj@email.com"
         icon={<FaEnvelope />}
-        error={errors.email?.message}
-        {...register("email")}
       />
       
       <FormInput
@@ -56,8 +33,6 @@ export const RegisterForm = () => {
         label="Hasło"
         placeholder="••••••••"
         icon={<FaLock />}
-        error={errors.password?.message}
-        {...register("password")}
       />
 
       <FormInput
@@ -65,13 +40,13 @@ export const RegisterForm = () => {
         label="Potwierdź hasło"
         placeholder="••••••••"
         icon={<FaLock />}
-        error={errors.confirmPassword?.message}
-        {...register("confirmPassword")}
       />
 
-      <Button type="submit" className="w-full" disabled={isSubmitting}>
-        {isSubmitting ? "Rejestracja..." : "Zarejestruj się"}
+      <Button type="submit" className="w-full">
+        Zarejestruj się
       </Button>
     </motion.form>
   );
 }; 
+
+export default RegisterForm;
