--- src/components/Auth/Forms/__tests__/LoginForm.test.tsx
@@ -1,1 +1,0 @@
-
